--- Pellets.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "collectibles.h"
-class Pellets :
-	public collectibles
-{
-	//normal pellets= increases score
-};
-
