--- SuperPellets.cpp
@@ -1,1 +1,0 @@
-#include "SuperPellets.h"
