--- SuperPellets.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "collectibles.h"
-class SuperPellets :
-	public collectibles
-{
-	//superpellets=superpowers
-};
-
