--- collectibles.cpp
@@ -1,1 +1,0 @@
-#include "collectibles.h"
