--- collectibles.h
@@ -1,6 +1,0 @@
-#pragma once
-class collectibles
-{
-	//abstract class for pellets and fruits and any future bonus features
-};
-
