+++ header-files/Pellets.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "collectibles.h"
+class Pellets :
+	public collectibles
+{
+	//normal pellets= increases score
+};
+
