+++ header-files/SuperPellets.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "collectibles.h"
+class SuperPellets :
+	public collectibles
+{
+	//superpellets=superpowers
+};
+
