+++ header-files/collectibles.h
@@ -1,0 +1,6 @@
+#pragma once
+class collectibles
+{
+	//abstract class for pellets and fruits and any future bonus features
+};
+
