--- p.txt
@@ -1,15 +1,0 @@
--1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 
--1 00 00 00 00 00 00 00 00 00 00 00 00 00 -1 
--1 00 -1 -1 -1 -1 -1 -1 00 00 00 00 00 00 -1
--1 00 -1 00 00 00 00 00 00 -1 -1 -1 -1 -1 -1
--1 00 -1 00 00 00 00 00 00 00 00 00 00 00 -1
--1 00 -1 00 -1 -1 -1 00 -1 -1 -1 00 00 00 -1
--1 00 00 00 -1 00 00 00 00 00 -1 00 -1 -1 -1
-00 00 00 00 -1 00 00 00 00 00 -1 00 00 00 00
--1 00 -1 00 -1 -1 -1 -1 -1 -1 -1 00 00 00 -1
--1 00 -1 00 00 00 00 00 00 00 00 00 00 00 -1
--1 00 -1 00 00 00 00 00 00 00 -1 -1 -1 00 -1
--1 00 -1 -1 -1 -1 -1 -1 -1 00 -1 00 00 00 -1
--1 00 -1 00 00 00 -1 00 00 00 00 00 00 00 -1
--1 00 00 00 00 00 00 00 00 00 00 00 00 00 -1
--1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1
