+++ source-files/collectibles.cpp
@@ -1,0 +1,1 @@
+#include "collectibles.h"
